feat: lint set up

+++ seeders/001-account.js
@@ -1,29 +1,26 @@
 module.exports = {
-  up: async (queryInterface) => {
-    const transaction = await queryInterface.sequelize.transaction();
-    await queryInterface.bulkInsert(
-      'account',
-      [{
-        email: "admin@mail.com",
-        password: "temp_password"
-      }],
-      { transaction },
-    );
+    up: async (queryInterface) => {
+        const transaction = await queryInterface.sequelize.transaction();
+        await queryInterface.bulkInsert(
+            'account',
+            [
+                {
+                    email: 'admin@mail.com',
+                    password: 'temp_password',
+                },
+            ],
+            { transaction },
+        );
 
-    await transaction.commit();
-  },
-  down: async (queryInterface, sequelize) => {
-    const { Op } = sequelize
+        await transaction.commit();
+    },
+    down: async (queryInterface, sequelize) => {
+        const { Op } = sequelize;
 
-    const transaction = await queryInterface.sequelize.transaction();
+        const transaction = await queryInterface.sequelize.transaction();
 
-    await queryInterface.bulkDelete(
-      'account', 
-      { email: 
-        { [Op.in]: ["admin@mail.com"] } 
-      }, 
-      { transaction });
+        await queryInterface.bulkDelete('account', { email: { [Op.in]: ['admin@mail.com'] } }, { transaction });
 
-    await transaction.commit();
-}
+        await transaction.commit();
+    },
 };
